requirements.txt
apturl==0.5.2
bcrypt==3.2.0
blinker==1.9.0
Brlapi==0.8.3
CacheControl==0.12.10
cachy==0.3.0
certifi==2020.6.20
chardet==4.0.0
cleo==0.8.1
click==8.3.3
clikit==0.6.2
colorama==0.4.4
command-not-found==0.3
crashtest==0.3.1
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.4
distro==1.7.0
distro-info===1.1build1
dnspython==2.8.0
duplicity==0.8.21
email-validator==2.3.0
fasteners==0.14.1
filelock==3.6.0
Flask==3.1.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.3.0
future==0.18.2
greenlet==3.5.0
gunicorn==25.3.0
gyp==0.1
html5lib==1.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.6
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==3.0.3
monotonic==1.6
more-itertools==8.10.0
msgpack==1.0.3
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
packaging==21.3
paramiko==2.9.3
pastel==0.2.1
pexpect==4.8.0
Pillow==9.0.1
pkginfo==1.8.2
platformdirs==2.5.1
poetry==1.1.12
poetry-core==1.0.7
protobuf==3.12.4
ptyprocess==0.7.0
pycairo==1.20.1
pycups==2.0.1
PyGObject==3.42.1
PyJWT==2.3.0
pylev==1.2.0
pymacaroons==0.13.0
PyMySQL==1.1.2
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu4.1
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
python-dotenv==1.2.2
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
requests-toolbelt==0.9.1
SecretStorage==3.3.1
shellingham==1.4.0
six==1.16.0
SQLAlchemy==2.0.49
systemd-python==234
tomlkit==0.9.2
typing_extensions==4.15.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
virtualenv==20.13.0+ds
wadllib==1.3.6
webencodings==0.5.1
Werkzeug==3.1.8
WTForms==3.2.1
xdg==5
xkit==0.0.0
zipp==1.0.0
